Split out app into two files, refactored a lot of code, created feature extraction + removal functions. Need to source function file into main app file at some point, visuals need work.
